made seperate file for notification
